Did some things
Added another frame of animation to the gunshot 'animation'.
Added sprites for checkpoints (they exist in the game files, but are not implemented yet.)
Added the charge meter (firing your gun consumes charge now).
Explicitly states the data type for a bunch of the exposed variables for the grounded enemies.
Lowered the trash eaters speed.
Some electrical wire improvements
and did some work on level 10